thrift/rust: allow serde to be overridden on a type-by-type basis
Summary:
Sometimes you want to DIY.

This adds a `(rust.serde = "false")` attribute to override the global
serde setting for specific user-defined types, so that a manual
implementation can be injected into the generated code.

This also supports `(rust.serde = "true")` if you want to add derived serde to a
specific type without enabling it globally.

#forcetdhashing

Reviewed By: dtolnay

Differential Revision: D34159451

fbshipit-source-id: 349479ae16708b670f646d2c383bd18addd6ec6a