Improve error message when deserializing malformed json
Summary:
Improve the error message that occurs when decoding malformed JSON by displaying the char values of the missing tokens.

Prior to the change, a missing `}` would appear as:
```
Expected [125], not enough remaining
```
which can easily be mistaken for the deserializer expecting 125 more bytes.  With this change the deserializer now displays the error:
```
Expected the following chars: "}", not enough bytes remaining
```

Reviewed By: dtolnay

Differential Revision: D34189681

fbshipit-source-id: 5fae6f8ec1b90f58a88d953e36a3fec486c8d031