thrift/go: Remove the go-remote generation
Summary:
Remove the "remote" binary generated with every thrift service.  This is similar to the intention of the `pyremote" to make a generated binary to interact with a service but it has never been used inside Meta (no buck build rules generated) and has never been updated to support standard/modern transports/protocols.

There should be some win here on every thrift service with go gen in both generation time as well as caching the outputs.

Reviewed By: codyohl, carlverge

Differential Revision: D33826002

fbshipit-source-id: ac188c5071eb88e2a6d2fa5f05a4e359253e9a8f