Add templates for generating metadata in py3lite
Summary:
Adds the template files which return a ThriftMetadata as specified in thrift/lib/thrift/metadata.thrift. These will get used in later diffs.

The recursive structure of the calls here is based off of the cpp2 metadata implementation, it makes sure that a ThriftMetadata return will have all the correct structs/enums/etc as required by whatever type is being inspected.

There currently appears to be a bug in thrift py3lite and thrift py3 where `type:typedef?` never resolves to true, so that branch in the templates is currently doing nothing, but this will be fixed later.

Reviewed By: pyamane

Differential Revision: D34067230

fbshipit-source-id: f03dab76e29ec22f86746a4b076e9df1406b18aa