Pass protocolID to PooledRequestChannel constructor
Summary:
The current PooledRequestChannel implementation can potentially block the initial request fetching protocolID from the underlying channel. It's better to just require the protocolId to be passed to PooledRequestChannel constructor.
Moreover the current protocolId caching logic was completely broken for BINARY_PROTOCOL (because protocolID is 0 in that case), resulting in all calls being blocking. So it's quite unlikely that any code is even using it with binary protocol right now :)

Reviewed By: ngoyal

Differential Revision: D34355891

fbshipit-source-id: b068c9b22f5ae1821a0d463802fc7bbacebc991d