Using UnboundedQueue for single bucket case
Summary:
We rely on AtomicNotificationQueue for Request queuing because it has a nice interface of telling us whether a queue is empty before pushed. The round robin nature guarantees that only one thread processes one bucket at a time (MPSC).

But in single bucket case, all the requests are pushed into a single queue, which make the situation a MPMC, we switch from AtomicNotificationQueue to UnboundedQueue to temporarily resolve this issue.

TODO:
We'll need to either add the enqueue-first mechanism to UnboundedQueue (ongoing), or implement that as an internal of RequestPile to get rid of this separate array of queues.

Reviewed By: roddym

Differential Revision: D34195654

fbshipit-source-id: ebcc83656b0875b059bab9df087d4342cb628f79