universal name (1/n): validate methods
Summary:
For universal name support in Python, just wrapping C++'s implementation is the most straightforward way. This diff wraps 3 methods:

* validate_universal_name
* validate_universal_hash
* validate_universal_hash_bytes

Once we started actually use them, we might found some of them not needed at all, we might also decide to change the wrapper API to be more Pythonic and easier to use.
But for now this should be a good start.

Reviewed By: Filip-F

Differential Revision: D34402500

fbshipit-source-id: eb3d78dd686d325538beb2d674664a6b53cb3fba