universal name (2/n): get methods
Summary:
For universal name support in Python, just wrapping C++'s implementation is the most straightforward way. This diff wraps 4 more methods:

* get_universal_hash_size
* get_universal_hash
* get_universal_hash_prefix
* maybe_get_universal_hash_prefix

Once we started actually use them, we might found some of them not needed at all, we might also decide to change the wrapper API to be more Pythonic and easier to use.
But for now this should be a good start.

Reviewed By: Filip-F

Differential Revision: D34402502

fbshipit-source-id: 1c5ee55c72b7151e777d3df7254f2f37bea1d6b4