universal names (3/n): matches_universal_hash
Summary:
For universal name support in Python, just wrapping C++'s implementation is the most straightforward way. This diff wraps 1 more methods:

* matches_universal_hash

Once we started actually use them, we might found some of them not needed at all, we might also decide to change the wrapper API to be more Pythonic and easier to use.
But for now this should be a good start.

Reviewed By: Filip-F

Differential Revision: D34402499

fbshipit-source-id: bff50e1550b0b7f233ac4cc22019f1b332786bc4