Remove direct cpp.template support.
Summary:
As some cpp.template annotations include templates with non-type template parameters, like small_vector.
To solve this, we will leave it up to the code gen to pick the right concrete c++ type for the given cpp.template usage (as we do today). I have left helpers in Adapt.h to help support adapted keys in maps and sets.

Reviewed By: Mizuchi

Differential Revision: D34252733

fbshipit-source-id: 9e2d68c34d82ae851d04018e5c33e831a3f93f18