Supporting queueObserver in RequestPile
Summary:
QueueObserver is currently installed in ThreadManager which reports important stats like queue lag counters

We are replicating this functionality in RequestPile as well.

Similarly to that in ThreadManager, the observer tracks the lag starting from when the request is pushed into the RequestPile, ending with when the request is pulled by the executor and ready for execution (the dequeue is always called by CPU threads).

Note:
In this diff we added RequestPileBase, which serves as a helper class that exposes functionalities that all the RequestPiles might need. The QueueObserver is installed inside RequestPileBase

Reviewed By: roddym

Differential Revision: D34090382

fbshipit-source-id: 49954c8b2445fda29ac000a32b66922b1b47a45f