Add new syntax to compiler
Summary:
Allows functions to return interactions.
Does not attempt to figure out whether a single return value or single first response value is an interaction or not during parse, assuming it is not and rectifying this in the standard mutator. Where the parse gives us this information, the interaction is placed in a new slot and validated in the mutator for simplicity, though that check could be moved to the standard mutator instead.
Since we need to support both old and new syntaxes and differentiate them in the codegen, the existing is_interaction_constructor method is not used for the new syntax and the returned_interaction method returns none for the old syntax.

Differential Revision: D34402736

fbshipit-source-id: 05f2ac2bdbd8575dc54d6d9f9ef9826dba3d47c5