Fix metadata path on imported types in lite_metadata
Summary:
Another minor mistake in the metadata templates, first `{{#type:need_module_path?}}` is always going to be true here since these aren't types files, so that is replaced with `{{#type:external_program?}}` which just checks the same thing but skips the types file check in the former.

Second the import prefix for the parent service is just wrong, it's a service, not a type. Oops.

Reviewed By: benhgreen

Differential Revision: D34490341

fbshipit-source-id: 272de458c526b26b553ec3027400dc34ab0c4178