write string fields into binaryValue of conformance::object
Summary: currently string fields are not validated to be utf8. This can cause problems in certain languages. For time being, we decided to save string fields in binaryValue of conformance::object. Schemaless serialization/deserialization does not work with json protocol because both binary & string is marked as T_STRING type in serailized data but both are encoded differently. Binary is base64 encoded and string is written as is. So during deserialization we cannot decide it correctly without schema.

Reviewed By: Alfus

Differential Revision: D34603821

fbshipit-source-id: 76debecba8f946321f9ced39073d50f756034f1c