Add support for multiple layers of standard mutators
Summary:
For mutators, the order of mutating matters as it modifies the thrift ast.

For example, we have `assign_uri` mutator that assigns uri for each thrift root definition.
For any mutator that needs to access uri, we need to make sure `uri` is assigned. Therefore, we will need to add an additional layer of mutator.

Reviewed By: Alfus

Differential Revision: D34587004

fbshipit-source-id: 580d1a2312159603ee6f51d599b84cca4daed1ff