let client constructor always accept (optional) cpp_transport
Summary:
The idea is, instead of controlling the rollout via compiler option, it's easier and safer to control it within the client factory.

In this diff, we let the constructor always accept an optional cpp_transport object. (note that asyncio doesn't work with cpp_transport yet, so it's a no-op).
In next diff I'll change the client behavior to use cpp_transport whenever available (passed in in constructor).
Finally I'll move the control knob to client factory.

Reviewed By: Filip-F

Differential Revision: D34728629

fbshipit-source-id: 89dd3d82d2adf670a8c6e67bd394a7bc4c04bfa8