Add StructPatch::patch
Summary:
StructPatch::patch() allows access to patches for individual fields. It is applies after any 'clear' operation.

If an 'assign' op is present when StructPatch::patch() is called, it is converted to clear() + field assign patches. The clear is needed, as an 'assign' patch completely replaces all existing fields, while field assign patches, would only assign to the fields known to the creator of the patch. If a new field is added, without the clear, it would erroneously remain unchanged after the patch.

Reviewed By: mariusae

Differential Revision: D34708317

fbshipit-source-id: 2cf3a1ca432aacdf35b11a46d7f5ca83d07c8bef