Add a parent class that both thrift py3lite Struct and Union inhreit from
Summary:
In py3 Union inherits from struct, so one can check for the presence of either Union or Struct by checking `isinstance(var, Struct)`, this breaks in py3lite, so to maintain compatibility, adding a parent class to both.

`types.pyi` is updated so that anyone checking for just Struct still gets access to all the shared methods.

Reviewed By: junzh0u

Differential Revision: D34154839

fbshipit-source-id: d94af0282d81e11d398fb91295eb5d8fb0b55500