support for inline docstrings
Summary:
A long-awaited feature, be able to place docstrings about thrift enums and fields inline. This is, because many of our files already contain inline comments explaining these fields. Moving the comments to a line above would reduce readability in these cases.

Fixing the compiler could be done in two ways: in order to have the `doctext` lex at the time of parsing, I would need to make that doctext be an active lexer element, part of the syntax. That might break any current expression not expecting this comment-like syntax everywhere. Or having it as part of the grammar.

As mentioned in the docstring (pun intended!) only inline ( = suffix) docstrings can be part of the grammar, not the current prefix ones.

Also renamed `parse_doctext()` and `capture_doctext()` to `push_doctext()` and `pop_doctext()` to be more indicative of how these behave.

Bonus point: doctext before the header element can be attached to the `program` , which is already a `t_node` and appear in the generated JSON. (see ThriftDoctext.json now)

Reviewed By: yfeldblum

Differential Revision: D32914595

fbshipit-source-id: 11c3b6ac4c0160a9394b1219dd072cdcb414aa28