Fix union behavior with wrapped fields
Summary:
Facebook :

Currently, the wrapped field in union is always "set" to the wrapper object and when the field is "unset", we unset the value stored within the wrapper.

This is invalid behavior for unions as at any time only one field should be "set".
This diff fixes this.

To set a wrapped field, we use the thrift internal method to create wrapper object and assign it to the field. This ensures that setter remains sync

#forcetdhashing

Reviewed By: yfeldblum

Differential Revision: D34636216

fbshipit-source-id: b3372fa81930969065b2efe33198e8b9cfc69754