Introduce a Statement entity to combine Definition and Header
Summary:
The new entity is `Statement`. The annotations are added on top of that in `StatementAnnotated`. Program now only contains `StatementList` with `StatementAnnotated` entities.

This avoids reduce-reduce conflicts when trying to add annotations to headers. With two lists or separate Header and Definition, the parser gets confused upon encountering @ - should it move to Header or Definition next.

A restriction of headers being specified before definitions is enforced by an extra validation logic in the parsing driver.

DestroyDocText is removed. It dates back to 2008 and doesn't seem necessary with current rules, as the doc context should be assigned to an appropriate entity as a part of StatementAttrs.

Reviewed By: Alfus

Differential Revision: D34184806

fbshipit-source-id: 2ebf03d5d55f341b1dd1bb5bc2e13b2305e672cc