emit only cpp2 handler errors to glog
Summary:
The current code also emits trusted exceptions to glog, such as application-overloadedd errors. The intention, though, is only to emit failures from the application code in the handlers, and only when the failures do not correspond to a user error declared in the `.thrift` file.

Note that the containing function is called by generated processor code only after testing for all user error types declared for the given service method in the `.thrift` file - see `{service}AsyncProcessor::throw_wrapped_{method}` generated functions.

Reviewed By: iahs

Differential Revision: D34795756

fbshipit-source-id: 658a11f99dca15a991abaeecf366ac5a27c86107