Fix TSimpleJSON handling of snake_case
Summary:
TSimpleJSON isn't handling thrift field with snake-case correctly.
The filed name used in the serialization is the mangled name and not the original name.

For instance a struct like this one:
```
struct SnakeProfile {
  1: i64 fb_id;
  2: string first_name;
  3: set<i64> friend_ids;
}
```
would be serialized like this:
```
{
  "fbId": 123,
  "firstName": "toto"
}
```
while it should (according to the spec) being serialized like this:
```
{
  "fb_id": 123,
  "first_name": "toto"
}
```

This fix make the serialization behave like intended but also make the deserialization backward compatible, i.e. an old JSON piece of data with `"fbId"` would be deserialize correctly.

Differential Revision: D34747633

fbshipit-source-id: 943f47b7a34d494d095c9e5899ec20c1642c6bf4