Removing duplicated fields in inherited classes
Summary:
Wrapper generated classes are defining the same private members in each child object and they do refer to the same object. If the object hierarchy is A<-B<-C (C is the root), all these objects define _headers parameter and refer to same HashMap instance.

The goal here is to define the _headers in C (root) only as protected and give access to B and A and eliminate duplication.

Reviewed By: avalonalex

Differential Revision: D34800771

fbshipit-source-id: ceb7ecb820279cf07c9ee94b9377e7536b4a0051