WeightedRequestPileQueue
Summary:
WeightedRequestPileQueue
This is a building block for weighted RequestPile.

This can be a replacement for what we are currently using in RoundRobinRequestPile by defaulting the weight to 1. The following diffs will do that.

It's an UnboundedQueue augmented with a control block which provides weighted constraints. Different control block plugins can provide different features (we by default provide 2 control blocks, one for 1D weights one for 2D).

When using as a MPSC queue, it also provides an API that tells the caller
whether an enqueue is the first enqueue before which the queue is empty.

Usage:
````
//
// This is a normal MPMC queue
WeightedRequestPileQueue<int> queue;
//
// This is a MPMC queue with control enabled
WeightedRequestPileQueue<int, true> queue;
// // weight is defaulted to 1
queue.enqueue(1);
// // set weight sum limit to 100
queue.setLimit(100);
// // enqueue an element with weight == 200
// // should reject
queue.enqueue(1, 200);
````

(Note: this ignores all push blocking failures!)

Reviewed By: roddym

Differential Revision: D34649713

fbshipit-source-id: ac5233d468b4793565e249a8766c634649d359e9