Add AsyncProcessorSet::getRequestsProcessor
Summary:
This method allows us to obtain the processor which can run the request so that we interact with it directly before we have gone through the AsyncProcessor interface boundary.

In the long term the goal is to make AsyncProcessorSet and AsyncProcessor interfaces distinct. The transport code will always deal with AsyncProcessorSet (and we'll likely only have one or two implementations of it). AsyncProcessor will continue to be the interface that is a customization point (with executeRequest() being the primary one for resource pools).

In the futue we may consider renaming this to RequestsProcessorSet and RequestsProcessor.

At the moment we need both interfaces in the same type to remain compatible with the non resource pools approach. In this context I have made AsyncProcessor the most derived subclass so that it's type which is widely used remains the same and documentation and code searches.

Reviewed By: ycking21

Differential Revision: D34598217

fbshipit-source-id: cdfbca4de16c231478d5dc0eb44142c6e4c79e70