Add GeneratePatch annotation
Summary:
This annotation triggers the patch_mutator to generate exactly the struct definitions deleted from the .thrift file.

Currently, only fields with base types are supported.

The exact delta between this and the original structs, in terms of code gen in each language, can be seen in the generated file of the patch test fixture. The main difference being, improved doc strings.

(Note: this ignores all push blocking failures!)

Reviewed By: mariusae

Differential Revision: D34741606

fbshipit-source-id: 8e27ad8f2f24194dbd0f75a316ae088f2a8be991