Refcount PooledRequestChannel::InteractionState
Summary: We currently delete the struct as soon as the interaction is terminated, which makes oneway semifuture methods in the interaction racy since you can't know how long you have to keep the interaction alive until it is done reading the id. To fix this we store a guard in each request lambda and defer the deletion until they are done.

Differential Revision: D34942548

fbshipit-source-id: ecb04d536873a36dd82c61dd05b759ac36c50220