Add AsyncProcessorHolder
Summary:
AsyncProcessorHolder is a simple implementation of something that provides the AsyncProcessorSet interface. In resource pools mode the transports (rocket, header, http) expect to be able to interact with something that implements AsyncProcessorSet. MultiplexAsyncProcessor does implement it but is overkill for the specialized connections (monitoring, debugging etc) where there is exactly one AsyncProcessor(Factory) involved in the connection. AsyncProcessorHolder provides the AsyncProcessorSet interface and contains a single AsyncProcessor.

(Note: this ignores all push blocking failures!)

Reviewed By: sethdelliott

Differential Revision: D34947865

fbshipit-source-id: 8d160fb46f3969410e30372ca0992a62d3a7a9bb