Skip serializing terse field if it is unset
Summary:
If the value of the field is equal to the intrinsic default, it is equivalent to unset, so it is eligible to elide serialization.

Note: will use `apache::thrift:op::isEmpty` instead after refactoring type/op library to have two separate buck rule that is used in codegen (C++14 compatible) and runtime (C++17 compatible).

Reviewed By: Mizuchi

Differential Revision: D35018481

fbshipit-source-id: 08fb8cb3d24a75cafdf185e849bf21b9f00fbc55