Add the fmt library to compiler/detail
Summary:
Add the fmt library to `compiler/detail` to be used for diagnostic formatting replacing the iostreams/lambda hack and for the floating point formatter with the round-trip guarantee.

For example, we could replace

```
ctx_.failure(*field, [&](auto& o) {
  o << "Field identifier " << field->get_key() << " for \""
    << field->get_name() << "\" has already been used.";
});
```

with a more readable

```
ctx_.failure(
  "Field identifier {} for \"{}\" has already been used.",
  field->get_key(), field->get_name());
```

The library is vendored like mustache to avoid adding an external dependency which would complicate usage.

In the future we will be able to replace this with C++20 `std::format`.

Reviewed By: iahs

Differential Revision: D35249646

fbshipit-source-id: ffef66e5bef656b711b3f453e47ff8d85dc96c56