Decouple the lexer from the parser
Summary:
The lexer shouldn't depend on the parser except for the token type (`yy::parser::symbol_type`). Fix this by moving the token construction logic from the parsing driver to the lexer where it belongs and adding the `lex_handler` interface for reporting out-of-band lexical constructs (currently only documentation comments).

Make reporting of lexer errors consistent while at it: now they are reported as failures and a special error token is returned which causes parser to stop.

Reviewed By: iahs

Differential Revision: D35230151

fbshipit-source-id: 257a4654f2fb987fbaeb4f54d567350738c744e3