Move diagnostic out of ast
Summary:
Move diagnostic out of ast because diagnostics have nothing to do with ast and are reported from all phases of translation including those that should not depend on ast conceptually such as the lexer.

It also makes the top-level compiler API cleaner because dealing with diagnostics is more common than with ast so we can avoid pulling in the unneeded ast interface in the common case. In practical terms, `thrift/compiler/compiler.h` no longer includes ast headers.

Decoupling of the lexer from the AST will be done in a follow-up diff.

Reviewed By: Mizuchi

Differential Revision: D35286093

fbshipit-source-id: a93397c69c9e2634b56482912293edd5646718ee