Update AsyncProcessor::executeRequest contract to exclude contextStack
Summary:
Before this change the caller of executeRequest was responsible for doing the request side of ctxStack interactions, while executeRequest implementation was responsible for doing the response side of the ctxStack interaction.
This was problematic for a few reasons:
1. It still didn't enable us to move the ctxStack completely outside of generated code.
2. It complicated executeRequest contract (having to pass ctxStack through its boundary), even though it's not the desired end state
3. It coupled the migration to the executeRequest API with the ctxStack contract changes

This change makes executeRequest responsible for both sides of the ctx stack interaction to address (2) and (3) above.

Reviewed By: roddym

Differential Revision: D35260434

fbshipit-source-id: 57eba1006982d83e3e86b68f3a0ecb692cf833d3