Add cpp.detail.adapted_alias and alias all Patch types
Summary:
This the final missing piece for completely replacing the thrift struct, with the patch wrapper type, by adding type alias for the 'public' name of the type, so the patch wrappers can be accessed directly by IDL name.

Ideally this would cause all generated type references to also use the alias, but that is more for aesthetics, and we can fix that in a subsequent diff or only in the structured annotation.

Reviewed By: Mizuchi

Differential Revision: D35622441

fbshipit-source-id: a76125e535779d1efb6f3fccf2385402fb940e0b