Fix type resolution for included programs
Summary:
Only the root program's scope is ever updated, so share it with included programs. This fixes some type resolution bugs that show up with mutators for included files.

While this maintains our current behavior, it likely is the cause of some resolution bugs, and makes it a lot harder to implement include aliasing, so we should ultimately fix it.

Reviewed By: thedavekwon

Differential Revision: D35739661

fbshipit-source-id: 531b9da5e7aa69e56aceb32b2141effe049b76b1