sort the included thrift modules by module path
Summary:
The included thrift modules are previously sorted by the abs file path (i.e. key of the std::map), and at different environements where the root directories are different included files, the order may be non-deterministic. Changing it to sort by the python module path instead so the order is not dependent on the environment/file system.

(Note: this ignores all push blocking failures!)

Reviewed By: junzh0u

Differential Revision: D35767794

fbshipit-source-id: 0132c9f6e1e5913e4cb0c242d1a4ea8e5c57b55e