fix richcmp implementations
Summary:
The original implementation took a shortcut of comparing the _fbthrift_data for structs, but it won't work in many cases (e.g. if a struct is embedded in a container or nested in another struct), due to the leading isset array in the _fbthrift_data. Two structs with exact same field values but different isset array won't be equal as well.

Fixing this in this diff.

Reviewed By: junzh0u

Differential Revision: D35704198

fbshipit-source-id: 1d799f9c481e6b1e40517544fc274b0203cf9516