Separate mstch_cpp2_interaction from mstch_cpp2_service
Summary:
This diff fixes a long-standing hack around interaction codegen.

Note that an interaction is actually also represented as a `mstch_cpp2_service` and `t_interaction` is a subclass of `t_service`. `service:parent_service_name` is supposed to refer to the outer-most service during interaction codegen. Today we set `service:parent_service_name` in the global options map so that recursing into `service:interactions` does not reset it. We can achieve the same thing by making `service:parent_service_name` polymorphic, which we do in this diff by introducing `mstch_cpp2_interaction`.

Note that since interactions are also services, the `service_generator` will be used if a custom `interaction_generator` is not provided (which we only do for `cpp2`). This ensures that codegen for other languages are unaffected by this change.

Reviewed By: iahs

Differential Revision: D35819563

fbshipit-source-id: 22fdda7a960a927c5955c20c473ece7c641c6c62