merge java code templates for java2 and swift
Summary:
the reason why java2 and swift have different templates initially is due to complicated circular dependency problem. Now that this problem is fixed and java2 and swift shares runtime, we can merge their templates so that features can be shared effectively to aid migration.

We also mark `ServiceClient` and `ServiceAsyncClient` as deprecated as they are the only additional files generated by swift and we are aiming at deprecating them.

Differential Revision: D35907450

fbshipit-source-id: 21ddbaa49eea18c6f038e15f4f08bdf652bf01e7