Rename AnyType to Type and make it portable
Summary:
This the canonical representation for a 'type' value, with a nice c++ wrapper class. Eventually we could automatically convert from an IDL type to it, for example:  `1: type.Type type = i16`. This is actually possible in c++, for example: `*foo.type() = type::i16_t{}`

The rep.thrift defs completely replace the internal impl details, in addition to making it portable.
I also upgraded the rep to be future compatible with parameterized IDL definitions, if we ever decide to support those ;-).

Reviewed By: vitaut

Differential Revision: D35833546

fbshipit-source-id: 63a4be5abc0105f232aec61eef6dc7bcdda8b85c