Remove lexer constructor and assignment operators
Summary: These are remnants from a long time ago when lexer owned the source to only allow moving but not copying. Now that the sources are owned by source_manager copying is fine so remove explicit constructor and assignment operators relying on implicitly-generated ones. As a result the lexer API is now ridiculously simple.

Reviewed By: Mizuchi

Differential Revision: D35988919

fbshipit-source-id: d322b166c18b399e7c64197735175869f14b18e2