Update TCppServer to properly reload certificates
Summary:
As part of S269900 that caused S269863, we noticed that TCppServer's setup/startup doesn't register anything to watch for certificate changes.

`ThriftServer` does this as part of `ThriftServer::serve` but `TCppServer` (`CppServerWrapper`) and `BaseController` do not call that, they do their own initialization.

Call `getSSLCallbackHandle` to register necessary components.

Reviewed By: sethdelliott

Differential Revision: D35970936

fbshipit-source-id: 129ef4158d28a71ef6a033fd56f86123614b1367