Add cpp.TriviallyRelocatable annotation
Summary:
This adds the ability to mark a `thrift struct` as relocatable by annotating it with `cpp.TriviallyRelocatable`. E.g.,

```
cpp.TriviallyRelocatable
struct MyType {
  1: i32 someField;
}
```

Marking a struct with this annotation will help folly containers detect that it is safe to use `memcpy` to move objects upon resize or when items shift within the container. Since this avoids the need to call `std::move` for each individual item, it can lead to substantial speedups.

Reviewed By: Mizuchi

Differential Revision: D35959049

fbshipit-source-id: d107468ac589723522fb486da164c5abd71204bc