Add Type rep
Summary:
This the canonical representation for a 'type' value, with a nice c++ wrapper class. Eventually we could automatically convert from an IDL type to it, for example:  1: type.Type type = i16
I also made the rep future compatible with parameterized IDL definitions, if we ever decide to support those.

Reviewed By: bahadird

Differential Revision: D36110901

fbshipit-source-id: 7cd99511bd733c2e25062dd297ff2a9b78577b37