Fix the thrift python metadata library for external usage
Summary: The way the metadata library works is that it expects the metadata defined on a thrift type to be of type `apache.thrift.metadata.lite_types`, but by prefixing it with `{{program:root_module_prefix}}` this broke the import externally, removing this makes the import correct. Also includes a minor fix in the metadata library.

Reviewed By: junzh0u

Differential Revision: D35944043

fbshipit-source-id: 5fa041de16f94645cbc0e77cb0c719840f44ec42