Update the logo and remove a blog link
Summary: Move the Thrift logo from `doc` to `website/static/img` replacing the Docusaurus logo. Remove a blog link from the top bar because the blog post it refers to (https://engineering.fb.com/2014/02/20/open-source/under-the-hood-building-and-open-sourcing-fbthrift/) is somewhat outdated. There is another link to blog in a less prominent location under More at the bottom of the page.

Reviewed By: avalonalex

Differential Revision: D36175206

fbshipit-source-id: fd866c87678a6123fbcde370bfccc4d437f92b20