Fix thrift.thrift URIs
Summary:
Only very rarely should the filename be included in the package. In this folder, likely only the language specific .thrift files (and maybe internal.thrift). Mostly because it would currently be overkill to add individual directories for them, though if they get large enough, that is entirely possible.

The only exception is languages that require a 1:1 file:package naming, e.g. py and go)

Reviewed By: Mizuchi

Differential Revision: D36187428

fbshipit-source-id: c0822548e4ed8fefe275b3f174bd99c30083b912