py: convert converter in Cython
Summary:
Normally we want to keep `py_legacy` utils in pure Python for xplat compatibility, but converter is an exception.
Because if you need a converter you're already using `thrift-python` or `thrift-py3` which is already base on Cython.

Converting it to Cython gives us 20~60% performance win, depending on the data pattern.

Reviewed By: nanshu

Differential Revision: D36329798

fbshipit-source-id: 4954916eeb054244d7e7dd1fda75744e0a32b122