python: optimize py3/py-legacy to python conversion performance
Summary:
this diff rewrite the converter with reflection-based method, instead of serialzie-then-deserialize

it shows big performance wins in most scenarios, but there is one certain scenario where big regression is observed:

**converting a struct that contains a long list, from py3 to python**

I believe it's because C++ is just much more performant when it comes to iterating, comparing to Cython/Python

I don't have an easy way to measure the perf in prod-like data so can't say for sure whether it's common. But given all other scenarios show major performance win, I think this is still worth shipping.

And of coz it would be even better if we can optimize the long-list performance. Any suggestion?

Reviewed By: nanshu

Differential Revision: D36354796

fbshipit-source-id: afb21d96805afd3f84cf300d4440d83b4c9fd37a