Add structured type schema representations
Summary:
This is intended to be the canonical serialized format for Thrift schema, as we require a means of passing around dynamic schema information.

Ideally, when we generate the docs for this, they will *be* the spec for Thrift schemas. Thus should include specs for at least:
 - schema changes
 - restrictions
 - features
 - etc

Or more generally, it should only include things that are needed for a Type to be useful at runtime (vs code gen time for example).

This will be used by the Thrift runtime reflection API, delos entity, yaml, Any, etc for interpreting dynamic Thrift values.

Note that, while these structs might also be used in an AST rep, but that is not the primary goal of these definitions. Specifically, the abstract syntax tree (AST) representation needs to be a ~pure representation of the source file
with as little 'interpretation' as possible. Then we perform semantic analysis to distill the result of interpreting the source file in to the representations you see in this file.

Any and all feedback appreciated. As minimalist as possible for now.

Reviewed By: mariusae, Mizuchi, thedavekwon, rahku, bahadird

Differential Revision: D36319266

fbshipit-source-id: 7df3225ac51ec9644c77aeeff5eb68ea4b589088