Add Field::customDefault to schema.thrift
Summary:
Using the new type::ValueId typedef as, depending on the stage of processing, the ideal representation for a value in the schema could be one of many types, for example:
 - Any: Serialized, schemed
 - Protocol Value/Object: Id-based, deserialized, partial schema
 - AST Const: Name-based, deserialized and schema-less.
Additionally, if a constant immutable value repeats, we would like to be able to only store the contents of that value once, and refer to those contents multiple times. To support all of these case, a value is never stored directly in the schema, but rather via a 'value id', that is used to identify a value in a context-defined external storage location.

Reviewed By: thedavekwon

Differential Revision: D36497537

fbshipit-source-id: dee4429e7a8f005ca0962c41362fbb932b663d8f