py3: override __dir__ on Struct
Summary:
There're so many existing codes that call `dir` on thrift structs to get list of fields.
That makes any change that attempts to add new method on it virtually impossible to land without breaking anything.
This diff overrides `__dir__` as attempt to mitigate that issue.

It won't help with those other call sites that calls `__dict__` directly, but one step at a time. :|

Reviewed By: nanshu

Differential Revision: D36580686

fbshipit-source-id: ce92208db4dc51d3e8d713fc51939e7541d0eeb0