Fix use after free in python server
Summary:
The `ContextStack`s created in these handlers can still be used after the handler method returns, which means that the `funcName` variable can pass out of scope and later lead to a use after free.

I'm remedying this by just allocating a full table of all the valid function names and then picking the one that matches, in order to always keep that memory valid as long as the python handler is still in memory. I don't know if there's some better way to store the required string in memory because it is a very simple string, but `ContextStack` takes a byte array so there's not much I can do in terms of c++ memory management here.

Reviewed By: nanshu

Differential Revision: D36389375

fbshipit-source-id: c001c08146b9fdaa0003ae275ec1c46a43581bf2