Combine typedefs and structs into one section in cpp codegen
Summary: In order to support using cpp.Adapter{} on a typedef of struct we have to topo sort these definitions as the adapter trait needs access to the full struct definition. This means we must interleave them, which in mstch requires combining them into one section. This is gross in the mstch model, but seems to be the best we can do

Reviewed By: Alfus

Differential Revision: D36644026

fbshipit-source-id: 5154d65d6be949207597c430374e3c07b2ffc101