Add declared and undeclared exceptions to the server aggregate counts
Summary:
Currently, there is no way to provide server wide context on all the exceptions that are handled by the service.

This adds two more counters along with their definition:
* thrift.declared_exceptions: Exceptions that are thrown in the handler that are expected by the IDL.
* thrift.undeclared_exceptions: Exceptions that are thrown in the handler that are not expected by the IDL.

To achieve this, we extend the RequestContext to include if the call path comes from an exception or not. This is used later on by the `ThriftRequestCore` which distinguishes if this request is coming from the exception handling routines in the Thrift Handler.

The idea here would be to provide enough context at the server level to track all the aggregate number of requests without relying on client data to provide a high fidelity overview of how well a server is performing just by looking at the server itself.

Differential Revision: D36527060

fbshipit-source-id: 4791df3db04c45f2f578049ee339d7c6ccade270