Implement hack.name annotation
Summary:
With more uses of hack and package namespaces, a lot of hack reserved keywords which were previously prefixed with php namespaces are now resulting in compilation errors, so add an annotation `hack.Name` to rename such symbols.

NOTE: if both are specified then the first preference is always given to the unstructured annotation.

Reviewed By: thedavekwon

Differential Revision: D36732090

fbshipit-source-id: 967559965078e21709d10dad5f145599ea278cde