Fix weird mid-line spaces
Summary:
I saw that some Thrift constructors had extra spaces at the end of their constructors. Traced it back to here.

We're adding indent before the closing paren of the constructor header, which would make sense if we had all the arguments on different lines... but we don't have that, so it just adds two spaces in the middle of the line.

This diff fixes it by removing the mid-line indent.

Reviewed By: vitaut

Differential Revision: D36937790

fbshipit-source-id: 01034833359b59aa2c227790daf8079693a62250